requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.2
numpy==1.22.2
packaging==23.1
Pillow==9.5.0
pytesseract==0.3.10
sqlparse==0.4.4
typing-extensions==4.6.3
tzdata==2023.3
